extracted out participantList from cohortDetails
